+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+### (2021-07-29) v0.9.3
+
+- Switched from `File#renameTo(File)` to the more robust
+  `Files.move(Path, Path, CopyOptions...)` alternative. (#14)
+
+- Add rolling support via `maxBackupCount`. (#14)
+
+- Stop policies after stream close. (#26)
+
 ### (2020-01-10) v0.9.2
 
 - Shutdown the default `ScheduledExecutorService` at JVM exit. (#12)